[#1484] Pre-warm domain session cookie on "more options" from reading page
This is an implementation of a fix that @alierak and @kareila theorized in 2015,
but which wasn't feasible at the time because `fetch` either didn't exist or
barely existed.

If you don't have a "domain session cookie" for a given journal, your first
request to that domain loops through some redirects to set up the cookie. Which
is fine unless that first request is a POST, because the redirects turn it into
a GET and you'll lose the request body.

"More options" from the reading page seems to be the only situation where that
can happen, so let's just make it honk its horn before it comes around the
curve.

`fetch()`'s `'no-cors'` mode is specifically designed to let JavaScript code
warm up various kinds of caches for content that it might not have permission to
interact with, so we're using it for its intended purpose.