[#2536] Fix background color for "fancy-select" menus on Windows
The background color of a `select` element is inherited by the menu it opens
when activated.

On most platform/browser combos, setting the background to "none" seems to make
both the menus and their text use the platform default styling, but in Chrome
and Edge on Windows, it unsets the background and doesn't affect the style of
the text. On dark themes like Gradation, that makes the menu white-on-white.

Since the select is invisible anyway via `opacity: 0`, I don't see any downside
to leaving out the `background: none` rule.