[#2587] Add click-to-zoom for images in entry/comment content
Previously, we added CSS that limits image sizes, to avoid reading page blowouts
when people post large images. This works as expected most of the time, but:

- It sometimes shrinks images it shouldn't, like vertical comic strips.
- Sometimes you just want to see the full size and don't mind a temporary blowout.

This commit handles those cases by adding a click-to-zoom behavior for images in
entry/comment contents. On desktop, a special cursor indicates when you can use
it. The new behavior only targets images that aren't inside links (since those
already have a click behavior), and has a special exemption for poll bar graphs.

This also moves the shrink CSS into a global fragment that gets pulled in with
need_res, so we don't need to maintain it in every layout.