Feature: bcrypted passwords
This implements using properly bcrypted passwords with unique salts, per
best practices. This change has a couple of sharp edges:

1) ATOM posting and digest-auth will break. Unclear how impactful this
is or what we should consider doing (if anything) about it.

2) Renames used to verify that the accounts being merged shared a
password. We think that having validated email accounts that match is
good enough, though.

3) Probably some bugs, I've only tested that I can still log in/out,
haven't gone wild testing every flow.

The implementation leverages our dversion system so that users can be on
the 'old style' passwords and we can upgrade them when we are confident
and ready to roll.