Support API keys for protocol auth
This provides a path forward for clients like Semagic that haven't been
updated in years, but we should still support them.

Given that we're moving to using better practice password storage, we
can't continue to use passwords. Instead of building 'app password'
functionality like some services have, we're going to be using our API
keys for now.

To use Semagic after we switch to the new password system, you will need
to generate an API key and use that as your password.