disable broken functionality for now
The change_journal_type console command won't work until we decide how to update the expected behavior, which has been to assign the password of the community's owner to the affected journal when changing it from a community to a personal journal. For now, the console command is disabled and all the relevant tests are skipped.

As for the t/atom-post.t file, it seems linked to AtomAPI authentication, which is a rarely used feature of the site. Skipping these tests for now but not sure if updating or removing is the best path forward.