[#2639] fix or disable broken tests and related functionality (#2642)
* fix syntax errors

* remove unneeded calls to old password method

t/console-changecommunityadmin.t: There was a test to make sure the password of a community was blank after transferring ownership, but communities haven't been allowed to have passwords in a very long time.

t/proto-post-edit-roundtrip.t: Use string variables and the check_password method instead.

t/console-reset.t: Use a known password and the check_password method instead.

* disable broken functionality for now

The change_journal_type console command won't work until we decide how to update the expected behavior, which has been to assign the password of the community's owner to the affected journal when changing it from a community to a personal journal. For now, the console command is disabled and all the relevant tests are skipped.

As for the t/atom-post.t file, it seems linked to AtomAPI authentication, which is a rarely used feature of the site. Skipping these tests for now but not sure if updating or removing is the best path forward.

* remove password-matching test from t/rename.t

Since it is no longer possible for us to test that two users have the same password, we no longer enforce this as a constraint when deciding if a rename is allowed. Now we only make sure that the authenticated email addresses on the two accounts match.

* tidy