Fix change_journal_type console command to assign a new password
This command had been temporarily disabled because when changing a community journal to a personal journal, the old behavior was to copy both the email and password of the named community admin over to the converted journal. But the previously used method for accessing that password was recently removed in favor of better password practices, so we needed a different solution for that aspect of the change.

The modifications here are copied in part from the reset_password console command, which uses LJ::rand_chars to generate a random string and assigns that as the value of the new password, then communicates it directly to the user via email with the expectation that it will immediately be changed.

The actions taken by the change_journal_type console command are executed by the reset_email user method, which sends a message asking the user to confirm the new email address for the journal. For the case where we are also changing the password, the message sent now asks the user to both confirm the email and change the password. In the case of converting to a community, the password is silently reset to the empty string.