Scour inline BML from make_preview (finish 1-1 talkpost_do conversion)
Hardcoding 'hotcolor', since spellcheck's on borrowed time anyhow. Otherwise
this didn't seem to be doing anything interesting.

As of this commit, everything seems to work, and the Talk controller acts as a
1-1 conversion of the old talkpost_do.bml. There's no tests for any of this
behavior, naturally, but I ran a fairly hellish manual testing matrix before
and after the conversion:

- Currently (logged out, site login, openid login)
- Posting as (self, anon, user (one-off), user (login), openid (one-off), openid (login))
- Performing (post, preview, preview-then-post, whiff-captcha-and-fix)

For the most part, it preserved any existing bugs and didn't add any new ones.
One exception:

- Fixes #2572?! 🙌🏼