LJ::T::P::i - refactor checks/munging into prepare_and_validate_comment
This moves all of LJ::Talk::Post::init's permissions / coherency / formatting
rules into a separate function, which is still long as hell but which at least
has a reasonable scope of responsibilities. It also grabs some rules that
belonged with these but had ended up in the callers instead of init, presumably
because nobody wanted to edit init.

While we're at it: Move ml strings to the new talkpost_do.tt.text. Bin the old
file and the dead strings.