LJ::T::P::i - Refactor user lookup into authenticate_user_and_mutate_form
The form mutation was always one of the most haunted-ass pieces of init. We're
still doing a little bit of it, but mostly just so talkform can tell you've
logged in.

All of the downstream things that need the old behavior are already getting
what they need from prepare_and_validate_comment, which fakes it reliably.

LJ::Talk::Post::init now does about as little as possible.