Completely destroy LJ::Talk::Post::init
Sleep tight, you hunk of garbage. 💀

This moves all of the integration and stuff that's specific to receiving
comments from a web form into the Talk controller, which is where it seems to
properly belong. That includes authenticate_user_and_mutate_form, since it's
really just an implementation detail of those forms.