Some maintenance work on the reply forms
A fair bit going on in here, but it's all miscellaneous.

- Removing spellcheck from the talkform, since I'm about to rebuild preview.

- If the form was submitted with a username and password and we need to display
the same form again, we now pre-fill the password so you don't need to re-type
it. This makes the new preview behavior feel much more humane. It felt a little
squirrelly at first, but it's going over HTTPS anyway, and the "someone goes to
your computer as soon as you stand up" threat model has bigger fish to fry.

    If you submitted a WRONG password, we don't pre-fill that. That's handled by
    authenticate_user_and_mutate_form.

- Now that talkpost_do isn't perma-locked to the jquery resource group, we can
display the new talkform on it. That means we don't have to sniff around at
$LJ::ACTIVE_RES_GROUP anymore! (The only reason we were doing that was because
the new talkform looked wack on talkpost_do.) Knocking that out of the QR too,
while I'm at it.

- Leaving some breadcrumbs behind for the next person, with the stuff I learned
recently about the various ID formats.

- Rearranging some stuff for legibility.

- Make talkform call LJ::get_remote instead of having $remote passed in. The
behavior of the "from" options interacts directly with the global login state,
so you'll never want to override $remote with something fake.

- Heyyyy, we don't need to deal with LJ::Talk::Post::init vandalizing the
usertype field anymore!!! (The new auth function does still munge usertype, but
it does it to accurately reflect how the login state has changed since the form
was submitted. Init would just straight-up lie to me.)