init_s2journal_js: check the remote for iconbrowser permissions
Everything that ever calls this function passes the exact same
`$remote && $remote->can_use_userpic_select` value for $opts{iconbrowser}, and I
don't think there's any use case for checking against a user OTHER than the
currently logged-in one. So we might as well just change that argument to mean
"the iconbrowser is usable here," and move the permissions check into the
convenience function.

(The call in the Talk controller was currently using stale info. I was about to
change it to check $real_remote, but on further thought why not just make the
function better.)