Fix drifting talkform in Drifting
In the s2foundation beta, where the talkform adopts some of the quick-reply's
styling, the form on the reply page gets pushed down below the sidebar. This
fixes that.

This layout is a great example of why everyone hates float-based layouts these
days. The fix is a kind of 😒 hack, but the alternative is basically open-heart
surgery on this thing so... *shrug*