Comment out default global Foundation rules for images
These rules make too many assumptions to be compatible with this site, and in
particular they interfere badly with user-provided content. As far as I can
tell, none of the current pages rely on these CSS rules.

If we were using a viable living framework here, hacking the core files like
this would be a bad plan. But Foundation 5 is extremely end-of-life, AKA we
bought it so we might as well break it. We're already carrying patches to fix
upstream bugs that will never be addressed, so hey, what's one more.