DW::Formats - New helper package for markup formats
This new package isn't used by anything yet, but it sort of serves as a map for
the following commits.

The goal of the coming changes is to introduce *named markup formats* on DW.
We'll use these formats whenever we need to display user-entered body text,
and we'll preserve the format name alongside the source as a permanent record of
how the user expected their markup to be interpreted when they wrote it. This
will give us the freedom to update and change our formatting behaviors as
needed, without vandalizing millions of historical posts or requiring
world-ending database migrations to update old formatting.

STORAGE: Entries and comments both already include an `editor` prop, which will
store the format ID for a given post. Other kinds of body content might be
handled differently as we expand format support. (For example, user bios are
singletons, so their format might be stored as a userprop.) Additionally, we'll
save a user's preferred format for a given content type and use it as the
default in editing UIs.

VERSIONING: Formats are versioned, and any significant change to markup behavior
is now expected to result in a new format name. To keep things manageable,
editing UIs will only offer a limited number of "active" formats at any given
time. (Exception: when editing old content, its current format will be available
in the UI even if it's obsolete.)

USAGE: We'll be using named formats in two ways:

- DISPLAY: Formats will determine the behavior of the HTML cleaner when
preparing content for display. They can affect the behavior of things like line
breaks, special non-standard HTML tags, `@mentions`, and more.
- UI: The currently selected format within an editing UI can potentially affect
its behavior and layout. (Whether to initialize or deinitialize the rich text
editor, whether plain-text editing buttons should emit HTML or Markdown, etc.)

The initial set of formats is a codification of the formats DW **already uses,**
which are currently selected by emergent interactions between the
`opt_preformatted` prop, the imported/syndicated status, Secret Glyphs
(`!markdown`), and the editor prop (currently only used in emailed comments).