Escape data attributes in _html_option
Looks like this was just forgotten when the feature was added. Only got noticed
when we started setting data attributes in the userpic menu to wrangle the icon
preview in reply forms.

I'm using the same $ehtml argument that sets escaping for values and text, since
I can't conceive of a situation where we'd want the one escaped but not the
other.