Tweak `@mention` scanning regexps to fix end-of-sentence bug
Prior to this patch, `@mentions` went Bad if they occurred before a
sentence-ending period -- mentions that used a shortcut site name (like `.ao3`)
just exploded, and other mentions would work but swallowed the sentence-ending
period.

This tweaks the regexps a bit:

- Cleanup: `\w` implies both `\d` and `_`, so they're not needed.
- Cleanup: `-` has special meaning depending on where it occurs within the
character class square brackets, so it's better to always escape it.
- Cleanup: hostnames can have hyphens.
- Bug fix: Final character of the matched @mention must not be a period.