[#1841] include priv packages on /admin/priv page
When granting privileges to a particular user, in addition to the existing dropdown of privileges, I added a second dropdown list of defined priv packages. I decided I liked the look of it better, and it was easier to implement than if I'd smashed the two sets of options into one menu as I had originally proposed. (It was also easier to implement than if I'd tried to do this before encapsulating the logic for granting privileges in the controller.)

This is just for ease of use when granting packages. All other package operations, such as seeing which privileges are included in a package, still require use of the priv_package console command.