Always display beta features in a consistent order
The list of available features on /beta gets ordered at random on every visit,
which is disorienting. Whenever there's 2+ betas, it's real hard not to notice.

That's not the intended behavior, but the controller was only sorting betas by
their "start time", which we don't appear to actually use. So let's fall back to
sorting by the beta's ID when the start times are identical (i.e. first
comparison returns 0); arbitrary, but at least it won't jump around like that.