Email posts: Enable new `post-format` header, default entries to Markdown
This new header lets you choose which format you want to use for emailed
comments or entries. Since email is sort of halfway between a UI and an API, it
allows two forms of format values:

- Humans typing directly into their mail client should ask for either `html` or
`markdown` (aliases to the current versions of Casual HTML and Markdown), so
they can use what they're familiar with from the web UI.
- Automated integrations (or humans with very particular tastes) should specify
a "real" format ID like `html_casual0`, so they can continue to work predictably
when formats change in the future.

And since there's now a way to opt out of it, emailed entries now default to
Markdown, same as emailed comments.