convert creditcard.bml, gco_notify.bml, pp_notify.bml
I used a different name for the template from creditcard.bml to make it more obvious what it actually did, which was not as important as the filename made it sound.

None of these are believed to be in active use on the site currently, and so haven't been tested apart from making sure the controller module compiled.