Add new "sticky" method to ajaxtip
It already has an "error" method that doesn't auto-dismiss on a timer, but that
still dismisses as soon as you mouse past it. To allow some kind of interaction
with the result of an ajaxtip request, you want the pop-up to stay on the screen
until you click away from it. So this lets it do that.