Foundation: Reduce chaos by copying .button styles to .submit
Sass's `@extend` directive works by adding that selector to EVERY place where
the extended selector is used. This includes places like `:not()`
pseudo-classes, and it can result in some very top-shelf chaos and dismay.

Foundation's separation of concerns is pretty poor, and .button leaks so many
places that it's a real bad candidate for `@extend`ing. So even though copying
all of the actual .button rules into .submit looks absolutely disgusting, it
stops some weird rampages before they start.

This is another thing that might have good arguments against it if we were using
a living framework, but staying on an EOLed one is a lot like squatting in a
condemned house, so it's kind of like why NOT cut a hole in the floor and
install a makeshift dumbwaiter to work around the fact that the dishwasher is on
a different floor than the stove.