Foundation: Repurpose a main nav hack to make new Tropos match old Tropos
This was working around the .submit thing I changed in the previous commit,
which it doesn't need to do anymore.

But we do still need something to make the Foundation main nav behavior match
the old main nav behavior:

- Default Foundation top bar highlights the active top-level link AND the
currently hovered sub-nav link using the hover color.
- The old designs make the active top-level link match the sub-nav background,
and only the currently hovered link gets the hover color.

So we can use most of this rule to make the active top-level link use the menu
body background instead of the hover background.