Add controller (and templates) for updating default format userprops
A page that wants to offer the chance to update your default (like comment/entry
success pages) can do so with the default_editor_form.tt template fragment.

To reduce clutter, this is a one-button interface rather than a real
choose-a-value setting. That means it should only be displayed after the user
does something that involved choosing a non-default format, so we know which
format we're offering to switch to.