Add JavaScript for changing default format without leaving the page
A basic little ajax submit script that hits the default_editor RPC endpoint.
The form fragment can bring the JS in automatically if you're building it into a
page from the beginning, but if the form gets added to a page via ajax (like
with the quicker-reply), the page needs to require the script ahead of time.