Talkform: Carry over viewing_thread value, if included in POST
Quick-reply uses this field to return you to the exact thread sub-view you were
on when you started replying. And the old previewform used to also respect that
when previewing FROM the QR, due to its mutant practice of copying every
unhandled form field into a hidden input.

Well, that broke when I killed previewform, but this brings it back, and ALSO
brings return-to-thread-view to "more options" posts for the first time.

It doesn't do anything about edit links or no-JS/newtab reply links, yet.