Add thread to reply/edit links, and ensure consistency when expanding
This lets us return you to the same thread view you were in after you finish
editing a comment (or replying, if you're a no-JS person; we already handle new
comments if you go through the quick-reply).

The deal with `destination_thread` and `js_expand_url` is this: the JS thread
expander works by requesting a _normal HTML page_ for the thread you're trying
to expand, and then finding the comment bodies in it and shimming them into
place in the current page. Without some extra bookkeeping, that results in some
reply/edit links that will redirect you to a foreign thread context if you use
them. So this commit tweaks the URL used by the JS thread expander to include a
note about where the comments will actually be landing, and then we respect that
request when building the edit and reply URLs.