make lastmodtime in FAQ preview look more like in LJ::Faq->load
When previewing changes to a FAQ, the time shown is the full date and time, which is a bit jarring since the final FAQ display just shows the date. So this fixes the TODO to make it look more similar.

(It's not identical because strftime doesn't have a format for 1st, 2nd, etc which is language specific. LJ::Faq->load uses MySQL's DATE_FORMAT when reading the date from the database, but since this is a preview page, we haven't saved it to the database yet.)