fix Sphinx comment jobs
When the comment construction path was recently refactored, we missed a step and ended up submitting requests for Sphinx to reindex edited comment text without submitting the ID for finding the comment in the database, which is an unhappy outcome for all involved. This updates the job arguments to be valid again.