Keep concat_res responses from blowing up on empty cachebuster param
The res_includes system looks for the most recently changed file in a given
clump to populate the `?v=12341341` query param. But if all of the files in the
clump exist only in the static build folder (because they were compiled from a
different format, for example), you can get an empty timestamp (`?v=`). Since
the regexp in the concat_res handler only trimmed the query params if the
timestamp had at least one digit, that would 404 the whole thing.

(This wasn't a problem in prod yet because bin/build-static.sh copies compiled
SCSS files into the htdocs source dir before sending them to build/static.)