/dev/formats: Add handler/template for listing formats info
The user-facing docs for formats will be in the faq, but we'll still need a sort
of escape hatch with full details for developers, API users, folks with odd
email posting scripts, or others who might have a reason to provide a raw format
ID instead of picking an active format from the drop-down. So let's generate
that page directly from the code so it's always up-to-date, and link to it from
a footnote in the faq page.

/dev/ seems like the correct parent path for this.