convert /extcss/index.bml to DW/Controller/CSSProxy.pm
I followed the example of DW::Controller::EmbeddedContent to figure out how to approach this. Once I got past the request setup changes, the rest of the code was basically identical from the original BML.

Has the world moved on from typepad.com shared stylesheets? I left the special treatment in, but it seemed outdated.