Fix link colors in `th` elements that occur in a `tbody`
This hack was to work around the fact that Foundation puts a background color on
the `thead` that matches the link color in the rest of the document. But `th`
cells can also occur in the `tbody` (to act as a header for a row instead of a
column), in which case we want the normal link color. So restrict the reversed
links to the `thead` where they're needed.