[#2755] Fix wonky pagination links in s2foundation
So tbh, I really wanted to shim in a call to the components/pagination.tt
template instead, and just replace that whole custom situation. But the scope of
that started ballooning out of control, so let's fix the immediate grossness the
easy way, and maybe we can come back to that after we've totally ripped out the
nos2foundation siteviews styles.

Using an `.action-box` without an `.inner` (for the tinted background) seems
weird, but it was already doing that, so that's one more question for another
day.