new error templates for suspended and purged users
To be used by journal-related pages when trying to display information for a user that has been deleted and purged or a user that has been suspended for misconduct.

This also adds a dw.request_status function for templates to use, so that the caller doesn't have to remember to set the correct status explicitly.

(I don't know if the long garbage comments are still best practice, but I saw they were present in views/error/404.tt, so I emulated that.)