Tests: Pin (some) @-mention tests to latest format versions
I just had one of my occasional visions of the future, is all.

We already have a format alias for "give me the newest Markdown" (to accommodate
email posts), so this commit updates the @-mentions tests to use that by
default.

This way if we switch Markdown processors to one that deliberately eats the
escaping backslash in front of punctuation (like @), we'll get a nice early
warning that we need to tweak something in the rendering path.