helper function _profile_ml for DW/Logic/ProfilePage.pm
The profile.bml page uses BML::set_language_scope( '/profile.bml' ) to allow locally scoped strings in other modules to resolve correctly. This hack won't work when we transfer the page to templates, so this change makes the language scope explicit for each local string, with one line to change when switching over to the new context.