CleanHTML: rely on DW::Formats for defaults/aliases
This was sort of an artifact of the order I built these in, but anyway: since
Formats has utilities for resolving aliases and such, the HTML cleaner only
really needs to care about the "real" format IDs.

This has no real effect on normal behavior today, but it makes things easier to
maintain, and it makes things behave more predictably in tests. (Notably, fixes
a test failure I recently introduced in t/cleaner-markdown.t by trying to pass
the cleaner an alias that was only defined in Formats.)