Site skins: Revise form control hacks for replies and multiform
A few recent additions that need adjustment, and a few things I never got quite
right on the previous passes.

- I previously styled the "multiform" (multi-comment actions) submit button like
a submit (big and red on Tropo), but on further thought, it's of secondary
importance and doesn't need a big flashing 🚨. Demote it to normal button.
- The multiform button had a frustrating little vertical misalignment with its
associated dropdown. Make the heights match, and line them up nice.
- The new markup controls were a little too fluffy on site skins due to
Foundation's massive button/dropdown styles, so shrink them to the same height
as the icon controls. (This also fixes the quote button not quite matching the
dropdown's height.)
- Make these hacked button styles look a little more consistent with most of the
site skin buttons. (I'd previously missed the little shiny edge at the top.)
- Add some extra side padding to the hacked "inline" dropdowns (icon select,
markup select, multi-comment actions) so that Foundation's fake dropdown
triangle doesn't overlap the text when the longest item is selected.