additional Logic/ProfilePage methods for templates
Many user methods return a list of userids that the template will try to interpret as a list reference, but not always successfully:

- List containing one element -> variable set to that element.
- List containing no elements -> variable set to the empty string.

Using the .list TT method fixes the first problem but not the second; a null result will still be the empty string and not an empty list.

The cleanest solution is to stub replacement methods in ProfilePage that return list references instead of lists. We were already using ProfilePage for the profile-specific userid lists such as not_mutually_watched_userids, which will also be updated to return list references in a later commit.