Site skin comments: Attempt to fix the wooz
Attempting to fix some vertigo issues reported with the new site skin comments,
working on a hypothesis that it involves some combination of line length and
spacing.

This shrinks the average line length of entries and top-level comments by about
ten or fifteen characters per line, at least with the default font (verdana).
It also adds some vertical spacing around the body of a comment.