remove ineffective editjournal feature action:saveunsuspend
The intent of this feature was to automatically open a new support request when a suspended entry was edited to remove offending content. This had no effect because $LJ::UNSUSPENSION_REQUEST_SPCATID was never set.

Most of the actual changes here are related to removing the unsuspend_supportid entryprop, intended to link the suspended entry and the support request.