also clear not_approved when suspending an account
The way I've set this up, the userprop has a value of 1 if the account needs to be reviewed, and a value of 2 if it has been reviewed and determined to contain spam. So anything with a value of 2 will go into a second queue for account suspension. Once the account is suspended, the userprop needs to be cleared to remove it from the suspension queue.