Icon browser: Visual and behavioral revisions
- No text means no text: Yank the keywords menu that appeared below the active
icon in no-text mode. This was there to support a workflow that the browser's
previous design also bent over backwards to enable (picking a specific keyword
for a multi-keyword icon while in no-text mode), but on investigation with some
of the userbase, it turned out maybe _no one actually wanted that workflow._

- Yes text means less text: Yank the "comment" text, which wastes a ton of
precious space and which (while interesting in its home context on the icons
page) isn't relevant to choosing which icon to use for this post.

- One-click and done: As best as I can tell, the icon browser's
two-clicks-to-finish interaction was only ever there to accommodate the "pick
specific keyword from no-text" workflow, which is now gone. So let's cut the
clicks in half: clicking an icon selects its first keyword, and clicking
a keyword in yes-text mode selects that keyword. This also lets us remove the
"Select" button entirely, which saves some space and kills off a source of
jump-around annoyance.

- Revised markup: we're now using normal button elements to represent the
clickable icon and keyword buttons, instead of mutant `<a>` elements with an
aria role set. This gives better tab and focus behavior in some browsers
(Firefox uses a better focus ring style for buttons than for mutant tabbables),
and ought to yield better accessibility. Required a mass of knock-on changes
everywhere in the JS and CSS, naturally.

- Focus previous icon: Previous design was to focus the search field when
summoning the icon browser (except it was unreliable so I fixed it). This
amends that, so that if you already selected an icon (either with the browser or
the menu), that icon's button gets focused when summoning the browser again.
That way, you can use tab to navigate nearby icons. AS A WILD BONUS, it turns
out that in some browsers (desktop FF and Safari, so far) this scrolls the
currently selected icon into view, so if you accidentally chose the wrong icon
when you wanted the one next to it, you're in perfect position to switch your
pick if you open the browser again.

- Minor filter tweaks: Now that we're doing one-click-and-done, the intermediate
"selected" state doesn't really mean anything (except when indicating the
current value of the actual form element). So I changed the search filter to not
change the "selected" state. For an all-keyboard flow, just tab out of the
search field to focus the desired icon instead of filtering down to exactly one
and hitting enter.

- Visual updates: Centering and aligning things a bit more nicely, and revising
the "small + text" mode to hopefully waste less space.